separate column menu from icon
