+++ packages/grid/_modules_/grid/components/columnHeaders/ColumnHeaderMenuIcon.tsx
@@ -1,24 +1,21 @@
 import * as React from 'react';
-// @ts-expect-error fixed in Material-UI v5, types definitions were added.
-import { unstable_useId as useId } from '@material-ui/core/utils';
 import IconButton from '@material-ui/core/IconButton';
 import { classnames } from '../../utils/classnames';
 import { GridApiContext } from '../GridApiContext';
 import { GridColDef } from '../../models/colDef/gridColDef';
-import { GridColumnHeaderMenu } from '../menu/columnMenu/GridColumnHeaderMenu';
 
 export interface ColumnHeaderMenuIconProps {
   column: GridColDef;
+  columnMenuId: string;
+  columnMenuButtonId: string;
   open: boolean;
+  iconButtonRef: React.RefObject<HTMLButtonElement>;
 }
 
 export const ColumnHeaderMenuIcon = React.memo((props: ColumnHeaderMenuIconProps) => {
-  const { column, open } = props;
+  const { column, open, columnMenuId, columnMenuButtonId, iconButtonRef } = props;
   const apiRef = React.useContext(GridApiContext);
-  const columnMenuId: string = useId();
-  const columnMenuButtonId: string = useId();
   const ColumnMenuIcon = apiRef!.current.components.ColumnMenuIcon!;
-  const iconButtonRef = React.useRef<HTMLButtonElement>(null);
 
   const handleMenuIconClick = React.useCallback(
     (event: React.MouseEvent<HTMLButtonElement>) => {
@@ -46,17 +43,6 @@ export const ColumnHeaderMenuIcon = React.memo((props: ColumnHeaderMenuIconProps
       >
         <ColumnMenuIcon fontSize="small" />
       </IconButton>
-      <GridColumnHeaderMenu
-        columnMenuId={columnMenuId}
-        columnMenuButtonId={columnMenuButtonId}
-        field={column.field}
-        open={open}
-        target={iconButtonRef.current}
-        ContentComponent={apiRef!.current.components.ColumnMenu}
-        contentComponentProps={{
-          ...apiRef!.current.componentsProps?.columnMenu,
-        }}
-      />
     </div>
   );
 });
