+++ packages/grid/_modules_/grid/components/columnHeaders/GridColumnHeaderItem.tsx
@@ -1,4 +1,6 @@
 import * as React from 'react';
+// @ts-expect-error fixed in Material-UI v5, types definitions were added.
+import { unstable_useId as useId } from '@material-ui/core/utils';
 import {
   GRID_COLUMN_HEADER_KEYDOWN,
   GRID_COLUMN_HEADER_CLICK,
@@ -26,6 +28,7 @@ import { GridColumnHeaderTitle } from './GridColumnHeaderTitle';
 import { GridColumnHeaderSeparator } from './GridColumnHeaderSeparator';
 import { ColumnHeaderMenuIcon } from './ColumnHeaderMenuIcon';
 import { ColumnHeaderFilterIcon } from './ColumnHeaderFilterIcon';
+import { GridColumnHeaderMenu } from '../menu/columnMenu/GridColumnHeaderMenu';
 
 interface GridColumnHeaderItemProps {
   colIndex: number;
@@ -59,6 +62,9 @@ export const GridColumnHeaderItem = React.memo(
   }: GridColumnHeaderItemProps) => {
     const apiRef = React.useContext(GridApiContext);
     const headerCellRef = React.useRef<HTMLDivElement>(null);
+    const columnMenuId: string = useId();
+    const columnMenuButtonId: string = useId();
+    const iconButtonRef = React.useRef<HTMLButtonElement>(null);
     const {
       disableColumnReorder,
       showColumnRightBorder,
@@ -150,7 +156,16 @@ export const GridColumnHeaderItem = React.memo(
         {!disableColumnFilter && <ColumnHeaderFilterIcon counter={filterItemsCounter} />}
       </React.Fragment>
     );
-    const columnMenuIconButton = <ColumnHeaderMenuIcon column={column} open={columnMenuOpen} />;
+
+    const columnMenuIconButton = (
+      <ColumnHeaderMenuIcon
+        column={column}
+        columnMenuId={columnMenuId}
+        columnMenuButtonId={columnMenuButtonId}
+        open={columnMenuOpen}
+        iconButtonRef={iconButtonRef}
+      />
+    );
 
     React.useLayoutEffect(() => {
       const columnMenuState = apiRef!.current.getState().columnMenu;
@@ -213,6 +228,17 @@ export const GridColumnHeaderItem = React.memo(
           resizing={isResizing}
           height={headerHeight}
           {...resizeEventHandlers}
+        />
+        <GridColumnHeaderMenu
+          columnMenuId={columnMenuId}
+          columnMenuButtonId={columnMenuButtonId}
+          field={column.field}
+          open={columnMenuOpen}
+          target={iconButtonRef.current}
+          ContentComponent={apiRef!.current.components.ColumnMenu}
+          contentComponentProps={{
+            ...apiRef!.current.componentsProps?.columnMenu,
+          }}
         />
       </div>
     );
